+++ src/v2/models/__init__.py
@@ -5,9 +5,13 @@ from src.v2.models.contracts import (
     V2HealthResponse,
     V2Stats,
 )
+from src.v2.models.enums import DisciplineV2, OrganizationTypeV2, RepositoryTypeV2
 from src.v2.models.errors import V2ErrorResponse, V2ErrorType, V2FieldError
 
 __all__ = [
+    "DisciplineV2",
+    "OrganizationTypeV2",
+    "RepositoryTypeV2",
     "V2ErrorResponse",
     "V2ErrorType",
     "V2ExtractResponse",